+++ web-common/src/components/alert-dialog/alert-dialog-portal.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
-	import { AlertDialog as AlertDialogPrimitive } from "bits-ui";
+  import { AlertDialog as AlertDialogPrimitive } from "bits-ui";
 
-	type $$Props = AlertDialogPrimitive.PortalProps;
+  // type $$Props = AlertDialogPrimitive.PortalProps;
 </script>
 
 <AlertDialogPrimitive.Portal {...$$restProps}>
-	<slot />
+  <slot />
 </AlertDialogPrimitive.Portal>
