+++ web-common/src/components/dialog-v2/dialog-portal.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
-	import { Dialog as DialogPrimitive } from "bits-ui";
+  import { Dialog as DialogPrimitive } from "bits-ui";
 
-	type $$Props = DialogPrimitive.PortalProps;
+  // type $$Props = DialogPrimitive.PortalProps;
 </script>
 
 <DialogPrimitive.Portal {...$$restProps}>
-	<slot />
+  <slot />
 </DialogPrimitive.Portal>
